+++ lib/removeWordPress.py
@@ -43,13 +43,15 @@ class removeWordPress():
         self.themes = [
             "/wp-content/themes/twentytwelve/",
             "/wp-content/themes/twentythirteen/",
-            "/wp-content/themes/twentyfourteen/"
+            "/wp-content/themes/twentyfourteen/",
+            "/wp-content/themes/twentyfifteen/"
         ]
         self.static_file = [
             "/wp-includes/images/crystal/license.txt",
-            "/wp-includes/images/crystal/license.txt",
+            "/wp-includes/js/swfupload/license.txt",
             "/wp-includes/js/plupload/license.txt",
             "/wp-includes/js/plupload/changelog.txt",
+            "/wp-includes/js/tinymce/skins/lightgray/fonts/readme.md",
             "/wp-includes/js/tinymce/license.txt",
             "/wp-includes/js/tinymce/plugins/spellchecker/changelog.txt",
             "/wp-includes/js/swfupload/license.txt",